Fixes in KP6/KP7. Add defend on KP7 and KP8

+++ KP7/src/matrix.c
@@ -62,7 +62,7 @@ void matrixPrintNormal(Matrix *matrix) {
         int curRowPiInd = matrix->CIP->data[i];
         int nextRowPiInd;
         if (i == (n - 1)) {
-            nextRowPiInd = (int)(vectorSize(matrix->PI) - 1);
+            nextRowPiInd = (vectorSize(matrix->PI) - 1);
         } else {
             nextRowPiInd = matrix->CIP->data[i + 1];
         }
